server/requirements.txt
absl-py==0.7.0
astor==0.7.1
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.16.1
Pillow==5.4.1
protobuf==3.6.1
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
